modified:   src/main/java/ifsp/edu/br/task_list/config/SecurityConfig.java 	modified:   src/main/java/ifsp/edu/br/task_list/controller/LoginController.java 	new file:   src/main/java/ifsp/edu/br/task_list/controller/RegisterController.java 	modified:   src/main/java/ifsp/edu/br/task_list/service/UsuarioService.java 	modified:   src/main/resources/data.sql 	modified:   src/main/resources/schema.sql 	modified:   src/main/resources/static/css/styles.css 	modified:   src/main/resources/templates/login.html 	new file:   src/main/resources/templates/register.html

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS Usuario (
-    id_usuario INT AUTO_INCREMENT PRIMARY KEY,
+    id_usuario BIGINT AUTO_INCREMENT PRIMARY KEY,
     username VARCHAR(45),
     email VARCHAR(45),
     password VARCHAR(45)
